Showing title image on MainActivity
Will show the title image on MainActivity - it will fill the entire
width of the screen whereas the height will match itself and will use
the scaleType fitXY so that it best fits the users device. We'll also
enable adjustViewBounds so that we can see the changes. Finally, we add
a content description for accessibility purposes.

Padding was also removed from activity_main.xml.