Set up custom constructors
Choice constructor takes a String for the text (The label to be
displayed for the next page) and the int for the nextPage. These will be
set to the appropriate member variables.

Page has two constructors - both take in the int imageId that will be
displayed on the page and the String text (Storyline).
The first Page constructor also takes in two Choices, choice1 and
choice2. As the last pages won't have these choices, they won't be
required in the constructor and instead will be set to null.

As checking null isn't always the best option, we also created a
boolean, mIsFinal, which is defaulted to false but will be set to true
if the final page constructor runs.

Also added the appropriate getter and setter methods for mIsFinal.

In Story, we set up each page using the appropriate Page constructor.