Now loading and using the page
Updated member variable formatting.
Created new member variable String mName.
mName will be set to the name passed in by the Intent.

Refactored name to mName.

Now running loadPage() from onCreate().

Rather than just displaying the text straight away, we'll now format the
String so we replace the %1$s in the Story text with the parameter,
mName.
The %1 stands for parameter 1 and $s stands for String.
Only after formatting the text will we set it now.