Made pages dynamic
Pages will now update accordingly to what page the user is on.

Created new mCurrentPage member variable.
Loading page from onCreate() at 0 (Default however, changes after button
press).

loadPage() takes in an int of choice (What page is being loaded?)
Refactored page to mCurrentPage.

Set up OnClickListeners for the options so that they will get the next
page and load it.
Note that for our last pages, this is going to cause an error because
we're trying to getChoice1() and getChoice2() but, these don't exist!
We'll work on that next.