Fixed final pages
Added check for isFinal().

If it is final, hide mChoice1 and set mChoice2 to "PLAY AGAIN". If
mChoice2 is tapped, finish the StoryActivity, returning the user to
MainActivity.

Added some commented code if you want to clear out the name text (It
remains filled in, following the Android Activity Lifecycle unless
someone has a really bad phone that runs out of memory after running our
app (That's a pretty bad phone)).

Oh, and this is the last commit! :scream: :boom: